+++ src/hooks/plugins/useRarible/types.ts
@@ -1,0 +1,11 @@
+export type RaribleNetwork = "rinkiby" | "eth";
+export type RaribleTokenType = "ERC721" | "ERC1155";
+export type RaribleTokenTypeAssetClass =
+  | "ETH"
+  | "ERC20"
+  | "ERC721"
+  | "ERC1155"
+  | "ERC721_LAZY"
+  | "ERC1155_LAZY";
+
+export const DEFAULT_RARIBLE_NETWORK = "eth";